requirements.txt
Flask
requests
pymongo
python-telegram-bot
cryptography